I hate Java with a burning passion
So for some reason, the Lord knows why, Java decided it would be an
absolutely wonderful idea to switch the X and Y coordinates around
in Math.atan2(). It should be x, y but it's instead y, x. I am this
close to switching to C++ and forcing the kids to learn how to use it.
On Jah